Fix SteamVR driver bug in the new motion smoothing beta
The driver for Vive implements both IVRServerDriverHost_004 and IVRServerDriverHost_005. Since the Vive driver is always included with SteamVR, it makes no sense to have more than one interface version. The v4 sends garbage and made the driver crash, so we're just ignoring that call when we get a string with a very small pointer, or a device class that is not a small positive number.